fix(webapp): fix CASA feedback
