+++ .github/workflows/deploy.yaml
@@ -45,8 +45,9 @@ jobs:
               run: npm ci
             - name: Build Webapp
               run: npm run ts-build && npm run -w @nangohq/webapp build
-            - name: Inject env.js integrity into dist/index.html
-              run: node scripts/inject-env-js-integrity.mjs packages/webapp/dist/index.html "${{ vars.API_DOMAIN }}"
+            - name: Inject API origin into dist/index.html
+              run: |
+                  sed -E -i 's#(<meta name="nango-api-origin" content=")[^"]*(")#\1${{ vars.API_DOMAIN }}\2#' packages/webapp/dist/index.html
             - name: configure aws credentials
               if: inputs.stage != 'replit'
               uses: aws-actions/configure-aws-credentials@v4
