+++ .github/workflows/preview-webapp.yml
@@ -84,8 +84,9 @@ jobs:
             - name: Build Webapp
               run: npm run ts-build && npm run -w @nangohq/webapp build
 
-            - name: Inject env.js integrity into dist/index.html
-              run: node scripts/inject-env-js-integrity.mjs packages/webapp/dist/index.html "${{ vars.API_DOMAIN }}"
+            - name: Inject API origin into dist/index.html
+              run: |
+                  sed -E -i 's#(<meta name="nango-api-origin" content=")[^"]*(")#\1${{ vars.API_DOMAIN }}\2#' packages/webapp/dist/index.html
 
             - name: Configure AWS credentials
               uses: aws-actions/configure-aws-credentials@7474bc4690e29a8392af63c5b98e7449536d5c3a # v4
