+++ packages/webapp/index.html
@@ -9,7 +9,7 @@
         <link rel="manifest" href="/manifest.json" />
         <title>Nango</title>
 
-        <script src="/env.js?hash=%VITE_HASH%"></script>
+        <meta name="nango-api-origin" content="" />
     </head>
 
     <body class="bg-nav-gradient">
