+++ packages/webapp/src/index.tsx
@@ -1,28 +1,13 @@
-import posthog from 'posthog-js';
-import React from 'react';
-import ReactDOM from 'react-dom/client';
+import { loadRuntimeEnv } from './utils/loadRuntimeEnv';
 
-import App from './app/App';
-import { Providers } from './app/providers';
-import { globalEnv } from './utils/env';
-
-import './index.css';
-
-if (globalEnv.publicPosthogKey) {
-    posthog.init(globalEnv.publicPosthogKey, {
-        api_host: globalEnv.publicPosthogHost,
-        mask_personal_data_properties: true,
-        session_recording: {
-            maskAllInputs: true
-        }
-    });
+async function bootstrap() {
+    try {
+        await loadRuntimeEnv();
+        await import('./main');
+    } catch (err) {
+        console.error('Failed to bootstrap Nango webapp', err);
+        document.body.innerHTML = '<p>Failed to load application configuration. Please refresh the page.</p>';
+    }
 }
 
-const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
-root.render(
-    <React.StrictMode>
-        <Providers>
-            <App />
-        </Providers>
-    </React.StrictMode>
-);
+void bootstrap();
