+++ packages/webapp/src/utils/env.ts
@@ -1,5 +1,3 @@
-// Runtime configuration snapshot. window._env is set by /env.js, loaded as a blocking <script>
-// in index.html before the React bundle runs. It carries API URLs, feature flags, and third-party
-// keys (PostHog, Sentry, Stripe, etc.) that vary per deployment.
-// See packages/server/lib/controllers/v1/getEnvJs.ts for how it's constructed.
+// Runtime configuration snapshot. window._env is set by /env.js before the app bundle runs.
+// See packages/webapp/src/utils/loadRuntimeEnv.ts and packages/server/lib/controllers/v1/getEnvJs.ts.
 export const globalEnv = { ...window._env };
